feat: upgrade to Spring Boot 3 (including Axon 4.8 and Camunda 7.20).
The changes are mostly due to the javax-->jakarta package rename, but there are some more things that needed to be changed for Spring Boot 3 to work. One of them is the upgrade to a newer version of flapdoodle embedded MongoDB with considerable changes to the API. Another one is that the hack for getting the spring.application.name property as a default value into our configuration property classes does not seem to work anymore, so it is replaced by a BeanPostProcessor.

Fixes #852